👼Renammed divine points to divnie favors
